requirements.txt
fastapi~=0.116.1
pydantic~=2.11.7
pandas~=2.3.2
openpyxl~=3.1.5
python-docx~=1.2.0
uvicorn~=0.35.0
Jinja2>=3.1,<4.0
google-genai~=1.36.0